Pull the advanced query builder into its own file.
